requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
werkzeug==2.3.7
gunicorn==21.2.0
